// src/api/generated/schemas/userResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */
import type { RoleResponse } from './roleResponse';

export interface UserResponse {
  publicId?: string;
  email?: string;
  name?: string;
  roles?: RoleResponse[];
  createdAt?: string;
}